backend/training_scripts/requirements.txt
# Durian Detection Training Requirements
# Install with: pip install -r requirements.txt

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# YOLOv8 - Ultralytics
ultralytics>=8.0.0

# Roboflow - Dataset management
roboflow>=1.1.0

# Environment variables
python-dotenv>=1.0.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# ONNX export (optional)
onnx>=1.14.0
onnxruntime>=1.15.0
